refactor(frontend): one landing page — fold agents page into landing feature, drop marketing clone
features/marketing was a ~90% clone of the newer features/landing-page
(same scaffolding, nav, screenshot frames). The unique content — the
agents DLTL setup page — moves into landing-page/agents-page.tsx reusing
the landing nav/frames/styles (agent CSS + responsive rules ported);
/download now redirects to /landing; features/marketing is deleted.
Also fixes two stale strings in DocsPage (test:e2e rename, dead CLI ref).
