+++ frontend/src/features/landing-page/landing.module.css
@@ -654,3 +654,114 @@
     min-height: 82px;
   }
 }
+
+/* agents page (ported from marketing.module.css) */
+.agentGrid {
+  display: grid;
+  grid-template-columns: minmax(0, 0.78fr) minmax(0, 1.22fr);
+  gap: 1rem;
+  width: min(1240px, calc(100% - 48px));
+  margin: 0 auto;
+  padding: 0 0 86px;
+}
+
+.agentHero {
+  width: min(1240px, calc(100% - 48px));
+  margin: 0 auto;
+  padding: 64px 0 42px;
+}
+
+.agentLead {
+  max-width: 820px;
+  color: rgba(244, 242, 238, 0.72);
+  font-size: clamp(1rem, 1.2vw, 1.15rem);
+  line-height: 1.55;
+}
+
+.agentPanel {
+  padding: 1.25rem;
+}
+
+.agentTitle {
+  max-width: 760px;
+  margin: 0.75rem 0 0.8rem;
+  font-size: clamp(2.45rem, 4.8vw, 5rem);
+  line-height: 0.98;
+  font-weight: 730;
+  letter-spacing: 0;
+}
+
+.checkItem {
+  display: flex;
+  align-items: flex-start;
+  gap: 0.7rem;
+  padding: 0.9rem;
+  background: #090909;
+  color: rgba(244, 242, 238, 0.7);
+  font-size: 0.9rem;
+  line-height: 1.55;
+}
+
+.checklist {
+  display: grid;
+  gap: 1px;
+  margin-top: 1rem;
+  border: 1px solid rgba(255, 255, 255, 0.11);
+  background: rgba(255, 255, 255, 0.1);
+}
+
+.compactBlock {
+  border: 1px solid rgba(255, 255, 255, 0.13);
+  border-radius: 8px;
+  background: #080808;
+  color: rgba(244, 242, 238, 0.78);
+  font-family: var(--font-geist-mono), ui-monospace, monospace;
+}
+
+.compactBlock {
+  margin: 1rem 0 0;
+  padding: 0.9rem;
+  overflow: auto;
+  white-space: pre-wrap;
+  font-size: 0.78rem;
+  line-height: 1.58;
+}
+
+.steps {
+  border: 1px solid rgba(255, 255, 255, 0.13);
+  border-radius: 8px;
+  background: #0a0a0a;
+}
+
+.steps {
+  overflow: hidden;
+}
+
+.stepsHeader {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 1rem;
+  min-height: 48px;
+  padding: 0 1rem;
+  border-bottom: 1px solid rgba(255, 255, 255, 0.12);
+}
+
+/* agents page responsive rules (ported) */
+@media (max-width: 980px) {
+  .agentHero,
+  .agentGrid {
+    width: min(100% - 32px, 720px);
+  }
+  .agentGrid {
+    grid-template-columns: 1fr;
+  }
+  .agentLead {
+    font-size: 1rem;
+  }
+}
+@media (max-width: 620px) {
+  .agentTitle {
+    font-size: clamp(2.35rem, 11vw, 3.6rem);
+  }
+}
